+++ pool/queue/queue.ml
@@ -87,6 +87,7 @@ let work_job
         }
       | Ok () -> { job_instance with status = Succeeded }
     in
+    let%lwt () = Notifier.job_reporter job_instance in
     update database_label job_instance)
   else (
     Logs.debug ~src (fun m ->
@@ -159,7 +160,10 @@ let create_schedule () =
   create "job_queue" interval periodic_fcn
 ;;
 
-let start = create_schedule %> Schedule.add_and_start
+let start =
+  Notifier.before_start ();
+  create_schedule %> Schedule.add_and_start
+;;
 
 let stop () =
   registered_jobs := [];